Moved role enums and maps to headers of associated class
